feat(view): define Total Data in App.tsx
