--- packages/view/src/ViewNodeTypes.temp.ts
@@ -1,7 +1,0 @@
-import type { NodeType } from "./NodeTypes.temp";
-
-export type ViewNode = {
-  node: NodeType;
-  xSeq: number;
-  ySeq: number;
-};
