+++ packages/view/src/components/Detail/Detail.tsx
@@ -1,4 +1,7 @@
-const Detail = () => {
+import type { GlobalProps } from "types/global";
+
+const Detail = ({ data }: GlobalProps) => {
+  console.log(data);
   return <>Detail</>;
 };
