+++ packages/view/src/components/Statistics/Statistics.tsx
@@ -1,4 +1,7 @@
-const Statistics = () => {
+import type { GlobalProps } from "types/global";
+
+const Statistics = ({ data }: GlobalProps) => {
+  console.log(data);
   return <div>Statistics</div>;
 };
