+++ packages/view/src/components/TemporalFilter/TemporalFilter.tsx
@@ -1,4 +1,7 @@
-const TemporalFilter = () => {
+import type { GlobalProps } from "types/global";
+
+const TemporalFilter = ({ data }: GlobalProps) => {
+  console.log(data);
   return <>TemporalFilter</>;
 };
