+++ packages/view/src/components/VerticalClusterList/VerticalClusterList.tsx
@@ -1,4 +1,7 @@
-const VerticalClusterList = () => {
+import type { GlobalProps } from "types/global";
+
+const VerticalClusterList = ({ data }: GlobalProps) => {
+  console.log(data);
   return <div>VerticalClusterList</div>;
 };
