+++ packages/view/src/types/ViewNodeTypes.temp.ts
@@ -1,0 +1,7 @@
+import type { NodeType } from "./NodeTypes.temp";
+
+export type ViewNode = {
+  node: NodeType;
+  xSeq: number;
+  ySeq: number;
+};
