+++ packages/view/src/types/global.d.ts
@@ -1,1 +1,5 @@
+import type { ClusterNode } from "*/NodeTypes.temp";
+
 type FixMe = any;
+
+type GlobalProps = { data: ClusterNode[] };
